nats-export-service/requirements.txt
Flask
nats-py